feat!: [#878] extract tracker_policy section in core config section
