+++ packages/torrent-repository/tests/entry/mod.rs
@@ -43,22 +43,22 @@ fn rw_lock_parking_lot() -> Torrent {
 
 #[fixture]
 fn policy_none() -> TrackerPolicy {
-    TrackerPolicy::new(false, 0, false)
+    TrackerPolicy::new(0, false, false)
 }
 
 #[fixture]
 fn policy_persist() -> TrackerPolicy {
-    TrackerPolicy::new(false, 0, true)
+    TrackerPolicy::new(0, true, false)
 }
 
 #[fixture]
 fn policy_remove() -> TrackerPolicy {
-    TrackerPolicy::new(true, 0, false)
+    TrackerPolicy::new(0, false, true)
 }
 
 #[fixture]
 fn policy_remove_persist() -> TrackerPolicy {
-    TrackerPolicy::new(true, 0, true)
+    TrackerPolicy::new(0, true, true)
 }
 
 pub enum Makes {
